requirements.txt
matplotlib==3.7.5
numpy==1.24.3
numpy==1.24.4
Pillow==11.0.0
streamlit==1.40.0
tensorflow==2.10.0
